add explicit test of whether the labels are int, its definitely platform dependent

+++ src/deepforest/dataset.py
@@ -119,19 +119,19 @@ class TreeDataset(Dataset):
                 # channels last
                 image = np.rollaxis(image, 2, 0)
                 image = torch.from_numpy(image).float()
-                targets = {"boxes": boxes, "labels": labels}
+                targets = {"boxes": boxes, "labels": labels.astype(np.int64)}
 
                 return self.image_names[idx], image, targets
 
             augmented = self.transform(image=image,
                                        bboxes=targets["boxes"],
-                                       category_ids=targets["labels"])
+                                       category_ids=targets["labels"].astype(np.int64))
             image = augmented["image"]
 
             boxes = np.array(augmented["bboxes"])
             boxes = torch.from_numpy(boxes).float()
             labels = np.array(augmented["category_ids"])
-            labels = torch.from_numpy(labels)
+            labels = torch.from_numpy(labels.astype(np.int64))
             targets = {"boxes": boxes, "labels": labels}
 
             return self.image_names[idx], image, targets
